+++ src/components/DatePicker/datePickerChangelogData.ts
@@ -19,6 +19,15 @@ export const changelogData: ChangelogData[] = [
     ],
   },
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: [
+      "Updates the `initialDate` and `initialDateTo` props to accept an empty string and adds optional `placeholder` and `placeholderTo` props.",
+    ],
+  },
+  {
     date: "2023-12-07",
     version: "2.1.3",
     type: "Update",